requirements.txt
Django==4.0
django-bootstrap5==21.3
django-admin-autocomplete-filter==0.7.1
Django-Verify-Email==1.0.9
